requirements.txt
# Image/Video Processing
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0

# AWS S3
boto3==1.34.0
python-dotenv==1.0.0
